Ollama-Project/requirements.txt
requests
faiss-cpu
sentence-transformers
python-dotenv
torch
transformers
diffusers
Pillow
datasets
soundfile
gradio
ollama
langchain
plotly
huggingface_hub
